requirements.txt
pandas
sqlalchemy
psycopg2-binary
scikit-learn
xgboost
shap
matplotlib
seaborn
jupyter
python-dotenv